增加删除和完成功能

+++ WorkSpace/Web/todolist/services/todo.js
@@ -24,15 +24,23 @@ Todo.findAll = function (callback) {
 }
 
 Todo.findByUserId = function (userid, callback) {
-    var sql = 'select * from 3014218084_todo where user_id = \'' + userid + '\'';
+    var sql = 'select * from 3014218084_todo where user_id = \'' + userid + '\''  + " && finished <> 1";
     db.query(sql,function (err,data) {
         callback(err,data);
     })
 }
 
+Todo.findByUserIdTime = function(user,callback){
+    var timeStart = user.endTime - (3600*24 -2);
+    var sql ='select * from 3014218084_todo where user_id = \'' + user.user_id + '\'' +' && start_time >=  \'' + timeStart + '\'' + ' &&  start_time <=\' '+ user.endTime + '\''+' && finished <> 1  order by set_time';
+    console.log(sql);
+    db.query(sql,function (err,data) {
+        callback(err,data);
+    });
+
+}
 
 Todo.add = function (todo,callback) {
-
     var sql = "insert into 3014218084_todo (user_id,set_time,start_time,content,estimated_time) values('"+todo.user_id+"','"+todo.set_time+"','"+todo.start_time+"','"+todo.content+"','"+todo.estimated_time+"')";
     db.query(sql,function (err,data) {
         callback(err,data);
@@ -40,13 +48,24 @@ Todo.add = function (todo,callback) {
 }
 
 
-Todo.delete = function (id,callback) {
-    var sql = "delete from 3014218084_todo where id =" + id;
+Todo.delete = function (ids,callback) {
+    var sql = "delete from 3014218084_todo where id =" + ids.id + " &&  user_id=" + ids.user_id;
+    console.log(sql);
     db.query(sql,function (err,data) {
         console.log(data);
         callback(err,data);
     })
 }
+
+Todo.patch = function (ids,callback) {
+    var sql = "update 3014218084_todo set finished=1 where id =" + ids.id + " &&  user_id=" + ids.user_id;
+    console.log(sql);
+    db.query(sql,function (err,data) {
+        console.log(data);
+        callback(err,data);
+    })
+}
+
 
 
 module.exports = Todo;
